config file separation
